% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIsimdata.R
\name{simulate_cat_Z1_Z2_null}
\alias{simulate_cat_Z1_Z2_null}
\title{Generate Data with Categorical Zs}
\usage{
simulate_cat_Z1_Z2_null(
  n = 1000,
  K1 = 5,
  K2 = 5,
  sigma_y = 1,
  sigma_x = 1,
  seed = NULL
)
}
\arguments{
\item{n}{Integer. Sample size.}

\item{K1}{Integer. Number of categories for Z1.}

\item{K2}{Integer. Number of categories for Z2.}

\item{sigma_y}{Numeric. Standard deviation of noise added to Y.}

\item{sigma_x}{Numeric. Standard deviation of noise added to X.}

\item{seed}{Integer or NULL. Random seed for reproducibility.}
}
\value{
A data frame with columns X, Y, Z1, and Z2.
}
\description{
Generates data with a hard case scenario where X and Y are influenced by two Z variables in a nonlinear manner.
}
